@@ -16,6 +16,7 @@ type BackBoneCollection struct {
 	CU169    ISPCollection
 	CU9929   ISPCollection
 	CM       ISPCollection
+	CMIN2    ISPCollection
 	EDU      ISPCollection
 	CST      ISPCollection
 }
@@ -23,15 +24,18 @@ type BackBoneCollection struct {
 type ISPCollection struct {
 	ISPName string
 	IP      string
+	IPv6    string
 }
 
 const (
 	CT163  string = "电信 163 AS4134"
 	CTCN2  string = "电信 CN2 AS4809"
 	CU169  string = "联通 169 AS4837"
-	CU9929 string = "联通 A网 AS9929"
-	CM     string = "移动 骨干网 AS9808"
+	CU9929 string = "联通 A网(CNC) AS9929"
+	CM     string = "移动 CMNET AS9808"
+	CMIN2  string = "移动 CMIN2 AS58807"
 	EDU    string = "教育网 CERNET AS4538"
+	CST    string = "科技网 CSTNET AS7497"
 )
 
 var TestIPsCollection = AllLocationCollection{
@@ -46,22 +50,48 @@ var Beijing = BackBoneCollection{
 	Location: "北京",
 	CT163: ISPCollection{
 		ISPName: CT163,
-		IP:      "106.37.67.1",
+		IP:      "ipv4.pek-4134.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.pek-4134.endpoint.nxtrace.org.",
+	},
+
+	CTCN2: ISPCollection{
+		ISPName: CTCN2,
+		IP:      "ipv4.pek-4809.endpoint.nxtrace.org.",
 	},
 
 	CU169: ISPCollection{
 		ISPName: CU169,
-		IP:      "123.125.96.156",
+		IP:      "ipv4.pek-4837.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.pek-4837.endpoint.nxtrace.org.",
+	},
+
+	CU9929: ISPCollection{
+		ISPName: CU9929,
+		IP:      "ipv4.pek-9929.endpoint.nxtrace.org.",
 	},
 
 	CM: ISPCollection{
 		ISPName: CM,
-		IP:      "211.136.25.153",
+		IP:      "ipv4.pek-9808.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.pek-9808.endpoint.nxtrace.org.",
+	},
+
+	CMIN2: ISPCollection{
+		ISPName: CMIN2,
+		IP:      "ipv4.pek-58807.endpoint.nxtrace.org.",
 	},
 
 	EDU: ISPCollection{
 		ISPName: EDU,
-		IP:      "101.6.15.130",
+		IP:      "ipv4.pek-4538.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.pek-4538.endpoint.nxtrace.org.",
+	},
+
+	// 中科院
+	CST: ISPCollection{
+		ISPName: CST,
+		IP:      "ipv4.pek-7497.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.pek-7497.endpoint.nxtrace.org.",
 	},
 }
 
@@ -69,32 +99,42 @@ var Shanghai = BackBoneCollection{
 	Location: "上海",
 	CT163: ISPCollection{
 		ISPName: CT163,
-		IP:      "101.226.28.198",
+		IP:      "ipv4.sha-4134.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.sha-4134.endpoint.nxtrace.org.",
 	},
 
 	CTCN2: ISPCollection{
 		ISPName: CTCN2,
-		IP:      "58.32.4.1",
+		IP:      "ipv4.sha-4809.endpoint.nxtrace.org.",
 	},
 
 	CU169: ISPCollection{
 		ISPName: CU169,
-		IP:      "139.226.206.150",
+		IP:      "ipv4.sha-4837.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.sha-4837.endpoint.nxtrace.org.",
 	},
 
 	CU9929: ISPCollection{
 		ISPName: CU9929,
-		IP:      "210.13.86.1",
+		IP:      "ipv4.sha-9929.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.sha-9929.endpoint.nxtrace.org.",
 	},
 
 	CM: ISPCollection{
 		ISPName: CM,
-		IP:      "120.204.34.85",
+		IP:      "ipv4.sha-9808.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.sha-9808.endpoint.nxtrace.org.",
+	},
+
+	CMIN2: ISPCollection{
+		ISPName: CMIN2,
+		IP:      "ipv4.sha-58807.endpoint.nxtrace.org.",
 	},
 
 	EDU: ISPCollection{
 		ISPName: EDU,
-		IP:      "202.120.58.155",
+		IP:      "ipv4.sha-4538.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.sha-4538.endpoint.nxtrace.org.",
 	},
 }
 
@@ -102,17 +142,42 @@ var Guangzhou = BackBoneCollection{
 	Location: "广州",
 	CT163: ISPCollection{
 		ISPName: CT163,
-		IP:      "106.37.67.1",
+		IP:      "ipv4.can-4134.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.can-4134.endpoint.nxtrace.org.",
+	},
+
+	CTCN2: ISPCollection{
+		ISPName: CTCN2,
+		IP:      "ipv4.can-4809.endpoint.nxtrace.org.",
 	},
 
 	CU169: ISPCollection{
 		ISPName: CU169,
-		IP:      "157.18.0.22",
+		IP:      "ipv4.can-4837.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.can-4837.endpoint.nxtrace.org.",
+	},
+
+	CU9929: ISPCollection{
+		ISPName: CU9929,
+		IP:      "ipv4.can-9929.endpoint.nxtrace.org.",
 	},
 
 	CM: ISPCollection{
 		ISPName: CM,
-		IP:      "120.198.26.254",
+		IP:      "ipv4.can-9808.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.can-9808.endpoint.nxtrace.org.",
+	},
+
+	CMIN2: ISPCollection{
+		ISPName: CMIN2,
+		IP:      "ipv4.can-58807.endpoint.nxtrace.org.",
+	},
+
+	// 中山大学
+	EDU: ISPCollection{
+		ISPName: EDU,
+		IP:      "ipv4.can-4538.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.can-4538.endpoint.nxtrace.org.",
 	},
 }
 
@@ -120,20 +185,24 @@ var Hangzhou = BackBoneCollection{
 	Location: "杭州",
 	CT163: ISPCollection{
 		ISPName: CT163,
-		IP:      "61.164.23.196",
+		IP:      "ipv4.hgh-4134.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.hgh-4134.endpoint.nxtrace.org.",
 	},
 	CU169: ISPCollection{
 		ISPName: CU169,
-		IP:      "60.12.244.1",
+		IP:      "ipv4.hgh-4837.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.hgh-4837.endpoint.nxtrace.org.",
 	},
 	CM: ISPCollection{
 		ISPName: CM,
-		IP:      "112.17.224.98",
+		IP:      "ipv4.hgh-9808.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.hgh-9808.endpoint.nxtrace.org.",
 	},
 	// 浙江大学 教育网
 	EDU: ISPCollection{
 		ISPName: EDU,
-		IP:      "210.32.2.1",
+		IP:      "ipv4.hgh-4538.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.hgh-4538.endpoint.nxtrace.org.",
 	},
 }
 
@@ -142,11 +211,12 @@ var Hefei = BackBoneCollection{
 	// 中国科学技术大学 教育网
 	EDU: ISPCollection{
 		ISPName: EDU,
-		IP:      "202.38.64.1",
+		IP:      "ipv4.hfe-4538.endpoint.nxtrace.org.",
+		IPv6:    "ipv6.hfe-4538.endpoint.nxtrace.org.",
 	},
 	// 中国科学技术大学 科技网
 	CST: ISPCollection{
-		ISPName: "中国科学技术大学 科技网 AS7497",
-		IP:      "210.72.22.2",
+		ISPName: CST,
+		IP:      "ipv4.hfe-7497.endpoint.nxtrace.org.",
 	},
 }
