@@ -1,12 +1,14 @@
 package ipgeo
 
 import (
+	"encoding/json"
 	"errors"
+	"strconv"
 	"sync"
 	"time"
 
+	"github.com/nxtrace/NTrace-core/wshandle"
 	"github.com/tidwall/gjson"
-	"github.com/xgadget-lab/nexttrace/wshandle"
 )
 
 /***
@@ -17,7 +19,7 @@ import (
  * 运作模型可以理解为一个 Node 一直在等待数据，当获得一个新的任务后，转交给下一个协程，不再关注这个 Node 的下一步处理过程，并且回到空闲状态继续等待新的任务
 ***/
 
-// IP 查询池 map - ip - ip channel
+// IPPool IP 查询池 map - ip - ip channel
 type IPPool struct {
 	pool    map[string]chan IPGeoData
 	poolMux sync.Mutex
@@ -46,33 +48,55 @@ func receiveParse() {
 		// json解析 -> data
 		res := gjson.Parse(data)
 		// 根据返回的IP信息，发送给对应等待回复的IP通道上
-		var domain string = res.Get("domain").String()
+		var domain = res.Get("domain").String()
 
 		if res.Get("domain").String() == "" {
 			domain = res.Get("owner").String()
 		}
 
+		m := make(map[string][]string)
+		err := json.Unmarshal([]byte(res.Get("router").String()), &m)
+		if err != nil {
+			// 此处是正常的，因为有些IP没有路由信息
+		}
+
+		lat, _ := strconv.ParseFloat(res.Get("lat").String(), 32)
+		lng, _ := strconv.ParseFloat(res.Get("lng").String(), 32)
+
 		IPPools.pool[gjson.Parse(data).Get("ip").String()] <- IPGeoData{
-			Asnumber: res.Get("asnumber").String(),
-			Country:  res.Get("country").String(),
-			Prov:     res.Get("prov").String(),
-			City:     res.Get("city").String(),
-			District: res.Get("district").String(),
-			Owner:    domain,
-			Isp:      res.Get("isp").String(),
+			Asnumber:  res.Get("asnumber").String(),
+			Country:   res.Get("country").String(),
+			CountryEn: res.Get("country_en").String(),
+			Prov:      res.Get("prov").String(),
+			ProvEn:    res.Get("prov_en").String(),
+			City:      res.Get("city").String(),
+			CityEn:    res.Get("city_en").String(),
+			District:  res.Get("district").String(),
+			Owner:     domain,
+			Lat:       lat,
+			Lng:       lng,
+			Isp:       res.Get("isp").String(),
+			Whois:     res.Get("whois").String(),
+			Prefix:    res.Get("prefix").String(),
+			Router:    m,
 		}
 	}
 }
 
-func LeoIP(ip string) (*IPGeoData, error) {
-	// 初始化通道 - 向池子里添加IP的Channel，返回IP数据是通过字典中对应键为IP的Channel来获取的
+func LeoIP(ip string, timeout time.Duration, lang string, maptrace bool) (*IPGeoData, error) {
+	// TODO: 根据lang的值请求中文/英文API
+	// TODO: 根据maptrace的值决定是否请求经纬度信息
+	if timeout < 5*time.Second {
+		timeout = 5 * time.Second
+	}
+
+	// 缓存中没有找到IP信息，需要请求API获取
 	IPPools.poolMux.Lock()
-	defer IPPools.poolMux.Unlock()
 	// 如果之前已经被别的协程初始化过了就不用初始化了
 	if IPPools.pool[ip] == nil {
 		IPPools.pool[ip] = make(chan IPGeoData)
 	}
-
+	IPPools.poolMux.Unlock()
 	// 发送请求
 	sendIPRequest(ip)
 	// 同步开启监听
@@ -83,10 +107,8 @@ func LeoIP(ip string) (*IPGeoData, error) {
 	case res := <-IPPools.pool[ip]:
 		return &res, nil
 	// 5秒后依旧没有接收到返回的IP数据，不再等待，超时异常处理
-	case <-time.After(5 * time.Second):
-		// default:
+	case <-time.After(timeout):
 		// 这里不可以返回一个 nil，否则在访问对象内部的键值的时候会报空指针的 Fatal Error
 		return &IPGeoData{}, errors.New("TimeOut")
 	}
-
 }
