@@ -6,8 +6,8 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/xgadget-lab/nexttrace/ipgeo"
-	"github.com/xgadget-lab/nexttrace/trace"
+	"github.com/nxtrace/NTrace-core/ipgeo"
+	"github.com/nxtrace/NTrace-core/trace"
 )
 
 type Reporter interface {
@@ -45,7 +45,7 @@ func experimentTag() {
 
 func (r *reporter) generateRouteReportNode(ip string, ipGeoData ipgeo.IPGeoData, ttl uint16) {
 
-	var success bool = true
+	var success = true
 
 	defer r.wg.Done()
 
@@ -114,11 +114,13 @@ func (r *reporter) InitialBaseData() Reporter {
 	r.targetTTL = uint16(len(r.routeResult.Hops))
 
 	for i := uint16(0); i < r.targetTTL; i++ {
-		traceHop := r.routeResult.Hops[i][0]
-		if traceHop.Success {
-			currentIP := traceHop.Address.String()
-			r.wg.Add(1)
-			go r.generateRouteReportNode(currentIP, *traceHop.Geo, i)
+		if i < uint16(len(r.routeResult.Hops)) && len(r.routeResult.Hops[i]) > 0 {
+			traceHop := r.routeResult.Hops[i][0]
+			if traceHop.Success {
+				currentIP := traceHop.Address.String()
+				r.wg.Add(1)
+				go r.generateRouteReportNode(currentIP, *traceHop.Geo, i)
+			}
 		}
 	}
 
